moves sidebar thunks into separate files
